Allows for seeks over 2G in bspatch.

+++ bspatch.h
@@ -41,7 +41,7 @@ extern "C" {
 struct bspatch_stream
 {
 	void* opaque;
-	int (*read)(const struct bspatch_stream* stream, void* buffer, int length, int type);
+	int (*read)(const struct bspatch_stream* stream, void* buffer, int64_t length, int type);
 };
 
 int bspatch(const uint8_t* source, int64_t sourcesize, uint8_t* target, int64_t targetsize, struct bspatch_stream* stream);
